Use "hessenberg" algorithm for calculating det
The default algorithm for calculating determinant is "df", which need
longer runtime and higher precision to get the accurate result. Refer
[trac 15712](https://trac.sagemath.org/ticket/15712) for more details.

- Bump the default maximum depth to 80.
- Modify the precision settings for computing and displaying.

+++ precision.py
@@ -1,5 +1,10 @@
 from sage.rings.integer import Integer
 from sage.rings.real_mpfr import RealField
-
-Real300 = RealField(Integer(10000))
-Real20 = RealField(Integer(20))
+# To enable the ability to find the 7th enery level at depth 60, we need at
+# least 350 bits with using "hessenberg" algorithm to calculate the
+# determinants.
+# Here are some values of the required bits:
+# - depth 60: 350 bits
+# - depth 80: 450 bits
+Real300 = RealField(Integer(500))
+Real20 = RealField(Integer(40))
